requirements.txt
# requirements.txt

Flask==2.2.5
python-dotenv==1.0.0
python-telegram-bot==13.15
ryanair-py==3.0.0
pandas==2.2.3
numpy==2.1.3
requests==2.31.0

